Is the USGS API offline? Handle it.
Upon successful connection to the API, store the response in the browser's local storage. Also store a time stamp with the data.

If the API is offline, and the user has visited the page before, the data is pulled from local storage and the earthquakes are still displayed.

Show the user the time stamp of the data. Notify the user when the API is offline, and when the data is pulled from the cached data.